Add Graph.DebugOperation
The method marks the start of a set of operations on the Graph, with
extra information opertionally provided in the second paramter. This
returns a function with a single End method to mark the end of the set
in the logs.

Refactor the existing graph Begin/End Operation calls to use this single
method.

Add calls to DebugOperation for each step while building the graph.